# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "MUXar (Musical UX Smart Enhancer)",
        "description": "REST API description",
        "contact": {
            "name": "Muxar dev team",
            "email": "muxar.wade@gmail.com",
            "url": "http://muxar.blogspot.it/"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        }
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/api/login": {
            "post": {
                "summary": "LogIn",
                "description": "Log in with GPlay account.",
                "operationId": "login",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "login details",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "login success"
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "summary": "LogOut",
                "description": "Log out from app.",
                "operationId": "logout",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "login success"
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "user": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
    }
}